requirements.txt
bcrypt==4.0.1
build==1.2.1
#CacheControl==0.14.0
#cachelib==0.13.0
#cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.1.0
cryptography==41.0.3
fastjsonschema==2.20.0
filelock==3.15.4
#flake8==6.0.0
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Mail==0.10.0
Flask-MySQL==1.5.2
flask-mysql-connector==1.1.0
Flask-MySQLdb==1.0.1
Flask-OAuthlib==0.9.6
Flask-WTF==1.1.1
fonttools==4.42.1
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.20.0
google-api-python-client==2.148.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-texttospeech==2.17.2
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.4
importlib_metadata==8.2.0
iniconfig==2.0.0
installer==0.7.0
ipykernel==6.25.1
ipython==8.14.0
isodate==0.6.1
itsdangerous==2.1.2
jaraco.classes==3.4.0
jedi==0.19.0
Jinja2==3.1.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keyring==24.3.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.3.0
msgpack==1.0.8
msrest==0.7.1
multidict==6.1.0
mysql-connector-python==9.0.0
mysqlclient==2.2.0
nest-asyncio==1.5.7
numpy==1.24.4
oauthlib==2.1.0
openai==1.51.1
opencv-python-headless==4.10.0.84
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
pkginfo==1.11.1
platformdirs==3.10.0
pluggy==1.5.0
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
prompt-toolkit==3.0.39
proto-plus==1.24.0
protobuf==5.28.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.10.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.9.1
pydub==0.25.1
pyflakes==3.0.1
Pygments==2.16.1
PyMySQL==1.1.0
pyparsing==3.0.9
pyproject_hooks==1.1.0
pytest==8.3.2
pytest-datafiles==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3
pyzmq==25.1.1
rapidfuzz==3.9.4
referencing==0.35.1
requests==2.31.0
requests-oauthlib==1.1.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.19.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
stack-data==0.6.2
tenacity==8.5.0
toml==0.10.2
tomlkit==0.13.0
toolz==0.12.1
tornado==6.3.2
tqdm==4.66.2
traitlets==5.9.0
trove-classifiers==2024.7.2
typing_extensions==4.12.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
virtualenv==20.26.3
wcwidth==0.2.6
Werkzeug==2.3.7
WTForms==3.0.1
xattr==1.1.0
yarg==0.1.9
yarl==1.13.1
zipp==3.19.2
